requirements.txt
python-qpid-proton
pyyaml
requests
pyinstaller